test: extract _Variants info-field unit tests to unit/variants/
